requirements/requirements_amd.txt
# YOLO Vision Studio - AMD GPU Requirements
# Optimized for DirectML-enabled AMD GPUs (RX 5700 XT and similar)

# Core ML Framework with DirectML
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
torch-directml>=0.2.5.dev240914

# YOLO Models and Computer Vision
ultralytics>=8.0.0
yolov5>=7.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Processing and Utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tensorboard>=2.13.0

# GUI and Interface
PyQt5>=5.15.0
PySide6>=6.5.0
tkinter-tooltip>=2.1.0
customtkinter>=5.2.0

# Screen Capture and Window Management
mss>=9.0.0
pygetwindow>=0.0.9
pyautogui>=0.9.54
Pillow-SIMD>=10.0.0.post1  # Faster image processing

# AMD GPU Support and Optimization
onnxruntime-directml>=1.15.0  # DirectML acceleration
onnx>=1.14.0
onnxruntime>=1.15.0

# Performance Optimization
numba>=0.57.0
psutil>=5.9.0

# Data Formats and I/O
PyYAML>=6.0
toml>=0.10.2
json5>=0.9.14
h5py>=3.9.0
pycocotools>=2.0.6

# Progress and Logging
tqdm>=4.65.0
rich>=13.0.0
loguru>=0.7.0
colorama>=0.4.6

# File and System Operations
watchdog>=3.0.0
pathlib2>=2.3.7
send2trash>=1.8.2

# Network and Web
requests>=2.31.0
urllib3>=2.0.4
websockets>=11.0.3

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Windows-specific (will be ignored on other platforms)
pywin32>=306; sys_platform == "win32"
wmi>=1.5.1; sys_platform == "win32"

# GPU Monitoring (AMD-compatible)
gpustat>=1.1.1
py3nvml>=0.2.7  # Alternative GPU monitoring

# Hyperparameter optimization
optuna>=3.3.0
hyperopt>=0.2.7

# Active learning and model interpretability
modAL>=0.4.1
lime>=0.2.0.1
shap>=0.42.0

# Database and caching
redis>=4.6.0
diskcache>=5.6.1

# Parallel processing
joblib>=1.3.0
multiprocessing-logging>=0.3.4

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Quality metrics and validation
scikit-metrics>=0.3.3
seqeval>=1.2.2

# DirectML-specific optimizations
# Note: These may require manual installation or specific versions
# protobuf>=3.20.0
# flatbuffers>=2.0.0 